fix:  Replace WinnerAlert with Alert, and extend its Redux implimentation to store color too. Also move single instance to ScreenTemplate
