+++ src/components/containers/PokerGame/PokerGame.container.tsx
@@ -11,14 +11,13 @@ import {
 import { Hand } from '@global/types';
 
 import { PokerTable } from '@components/library/PokerTable';
-import { DropDown } from '@components/library/DropDown';
 import { PokerCard } from '@components/library/PokerCard';
 import { PokerPlayer } from '@components/library/PokerPlayer';
-import { WinnerAlert } from '@components/library/WinnerAlert';
 
 import { useStoreActions, useStoreState } from '@redux/index';
 import { Button } from '@components/library/Button';
-import { ButtonColors } from '@components/library/Button/Button.styles';
+import { ColorOptions } from '@global/theme';
+import { Dropdown } from '@components/library/Dropdown';
 
 const GameContainer: FC = () => {
   const [numberOfPlayers, setNumberOfPlayers] = useState(0);
@@ -66,7 +65,7 @@ const GameContainer: FC = () => {
     if (!message || !winners) return;
 
     storeWinners(winners);
-    configureAlert({ isVisible: true, message });
+    configureAlert({ isVisible: true, message, color: ColorOptions.Success });
   }, [configureAlert, pokerHands, storeWinners]);
 
   const getCardVisibility = (index: number, playerNumber: number) => {
@@ -95,25 +94,12 @@ const GameContainer: FC = () => {
     [],
   );
 
-  const { isVisible, message } = useStoreState(
-    (state) => state.alert.alertConfig,
-  );
-
-  const handleOnModalClose = useCallback(() => {
-    configureAlert({ isVisible: false, message: '' });
-  }, [configureAlert]);
-
   return (
     <PokerGame>
-      <WinnerAlert
-        isVisible={isVisible}
-        handleOnModalClose={handleOnModalClose}
-        message={message}
-      />
       <>
         {pokerHands.length === 0 && (
           <PlayerSelectionFlexWrapper>
-            <DropDown
+            <Dropdown
               title={
                 numberOfPlayers > 0
                   ? `${numberOfPlayers}`
@@ -123,11 +109,7 @@ const GameContainer: FC = () => {
               makeSelection={storeNumberOfPlayers}
               activeItem={numberOfPlayers}
             />
-            <Button
-              borderColor={ButtonColors.Red}
-              disabled={numberOfPlayers === 0}
-              onClick={startGame}
-            >
+            <Button disabled={numberOfPlayers === 0} onClick={startGame}>
               Start game
             </Button>
           </PlayerSelectionFlexWrapper>
@@ -152,7 +134,6 @@ const GameContainer: FC = () => {
           <BottomActionButtonsContainer>
             {pokerHands.length > 0 && (
               <Button
-                borderColor={ButtonColors.Green}
                 onClick={handleDetermineWinner}
                 disabled={winners.length > 0}
               >
@@ -160,11 +141,7 @@ const GameContainer: FC = () => {
               </Button>
             )}
             {pokerHands.length > 0 && (
-              <Button
-                color={ButtonColors.Green}
-                onClick={handleReplay}
-                disabled={winners.length === 0}
-              >
+              <Button onClick={handleReplay} disabled={winners.length === 0}>
                 Replay
               </Button>
             )}
