+++ src/components/containers/PokerGame/state/alert-model.ts
@@ -2,7 +2,7 @@ import { action } from 'easy-peasy';
 import type { AlertModel } from '@redux/types';
 
 const alertsModel: AlertModel = {
-  alertConfig: { isVisible: false, message: '' },
+  alertConfig: { isVisible: false, message: '', color: undefined },
   configureAlert: action((state, payload) => {
     state.alertConfig = payload;
   }),
