+++ src/components/library/Alert/index.ts
@@ -1,0 +1,1 @@
+export { default as Alert } from './Alert.library';
