--- src/components/library/WinnerAlert/index.ts
@@ -1,1 +1,0 @@
-export { default as WinnerAlert } from './WinnerAlert.library';
