+++ src/components/templates/ScreenTemplate/ScreenTemplate.template.tsx
@@ -1,5 +1,7 @@
-import { memo, ReactNode, FC, CSSProperties } from 'react';
+import { memo, ReactNode, FC, CSSProperties, useCallback } from 'react';
 import { StyledScreenTemplate } from './ScreenTemplate.styles';
+import { Alert } from '@components/library/Alert';
+import { useStoreState, useStoreActions } from '@redux/index';
 
 interface ScreenWrapperProps {
   children: ReactNode;
@@ -10,8 +12,26 @@ const ScreenTemplate: FC<ScreenWrapperProps> = ({
   children,
   containerStyle,
 }) => {
+  const { isVisible, message, color } = useStoreState(
+    (state) => state.alert.alertConfig,
+  );
+
+  const configureAlert = useStoreActions(
+    (actions) => actions.alert.configureAlert,
+  );
+
+  const handleOnModalClose = useCallback(() => {
+    configureAlert({ isVisible: false, message: '' });
+  }, [configureAlert]);
+
   return (
     <StyledScreenTemplate style={containerStyle} fluid>
+      <Alert
+        isVisible={isVisible}
+        handleOnModalClose={handleOnModalClose}
+        message={message}
+        color={color}
+      />
       {children}
     </StyledScreenTemplate>
   );
